requirements.txt
# Hệ Thống Tự Động Hóa Tra Cứu và Tổng Hợp Thông Tin BHXH
# Python Dependencies - PRODUCTION READY
# Phiên bản: 2.0.0 - Complete Edition

# =============================================================================
# CORE DEPENDENCIES - Dependencies cốt lõi
# =============================================================================

# Excel processing - Xử lý Excel
openpyxl==3.1.2
xlsxwriter==3.1.9
pandas==2.1.4

# HTTP requests - Gửi HTTP requests
requests==2.31.0
urllib3==2.0.7
httpx==0.28.1

# Data processing - Xử lý dữ liệu
json5==0.9.14
python-dateutil==2.8.2

# =============================================================================
# GUI AND CONSOLE - Giao diện và console
# =============================================================================

# GUI dependencies - Dependencies cho GUI
tkinter  # Built-in với Python
Pillow==10.1.0

# Console output - Hiển thị console đẹp
rich==13.7.0
click==8.1.7

# =============================================================================
# LOGGING AND UTILITIES - Logging và tiện ích
# =============================================================================

# Environment variables - Biến môi trường
python-dotenv==1.0.0

# Logging - Ghi log
coloredlogs==15.0.1
structlog==23.2.0

# Progress bars - Thanh tiến trình
tqdm==4.66.1

# =============================================================================
# CONFIGURATION MANAGEMENT - Quản lý cấu hình
# =============================================================================

# Configuration files - File cấu hình
pyaml==6.0.1
toml==0.10.2

# =============================================================================
# ERROR HANDLING AND RETRY - Xử lý lỗi và retry
# =============================================================================

# Retry mechanisms - Cơ chế retry
retrying==1.3.4
tenacity==8.2.3

# =============================================================================
# DATA VALIDATION - Xác thực dữ liệu
# =============================================================================

# Data validation - Xác thực dữ liệu
cerberus==1.3.5
pydantic==2.5.0

# =============================================================================
# TEXT PROCESSING - Xử lý văn bản
# =============================================================================

# Text processing - Xử lý văn bản
re2==1.0.7
unidecode==1.3.7

# =============================================================================
# SYSTEM UTILITIES - Tiện ích hệ thống
# =============================================================================

# System monitoring - Giám sát hệ thống
psutil==5.9.6

# =============================================================================
# CHECK CCCD MODULE - Module Check CCCD
# =============================================================================

# Web API framework - Framework API web
fastapi==0.116.1
uvicorn[standard]==0.35.0

# HTML parsing - Phân tích HTML
beautifulsoup4==4.13.5
lxml==6.0.1

# Testing - Kiểm thử
pytest==8.4.1
respx==0.22.0
pytest-httpx==0.35.0

# =============================================================================
# BHXH TOOL DEPENDENCIES - Dependencies cho BHXH Tool
# =============================================================================

# Caching - Cache
cachetools==5.3.2
diskcache==5.6.3

# Async support - Hỗ trợ async
asyncio-throttle==1.0.2

# =============================================================================
# OPTIONAL DEPENDENCIES - Dependencies tùy chọn
# =============================================================================

# Subprocess (chỉ cho Python < 3.3)
subprocess32==3.5.4; python_version < '3.3'

# =============================================================================
# DEVELOPMENT DEPENDENCIES - Dependencies phát triển (tùy chọn)
# =============================================================================

# Code formatting - Định dạng code
# black==23.0.0

# Linting - Kiểm tra code
# flake8==6.0.0

# Type checking - Kiểm tra kiểu
# mypy==1.5.0